+++ server/src/main/kotlin/com/kamelia/hedera/plugins/ForceChangePassword.kt
@@ -3,7 +3,7 @@ package com.kamelia.hedera.plugins
 import com.kamelia.hedera.core.Errors
 import com.kamelia.hedera.core.response.MessageKeyDTO
 import com.kamelia.hedera.core.response.Response
-import com.kamelia.hedera.rest.auth.UserState
+import com.kamelia.hedera.core.auth.UserState
 import io.ktor.http.*
 import io.ktor.server.application.*
 import io.ktor.server.request.*
