+++ server/src/main/kotlin/com/kamelia/hedera/rest/user/Service.kt
@@ -10,12 +10,9 @@ import com.kamelia.hedera.core.ValidationScope
 import com.kamelia.hedera.core.constant.Actions
 import com.kamelia.hedera.core.response.ActionResponse
 import com.kamelia.hedera.core.response.MessageDTO
-import com.kamelia.hedera.core.response.MessageKeyDTO
 import com.kamelia.hedera.core.response.Response
-import com.kamelia.hedera.core.response.asMessage
 import com.kamelia.hedera.core.validate
 import com.kamelia.hedera.database.Connection
-import com.kamelia.hedera.rest.auth.SessionManager
 import com.kamelia.hedera.rest.configuration.DiskQuotaPolicy
 import com.kamelia.hedera.rest.configuration.GlobalConfigurationService
 import com.kamelia.hedera.rest.core.pageable.PageDTO
