+++ server/src/main/kotlin/com/kamelia/hedera/websocket/Handlers.kt
@@ -1,12 +1,10 @@
 package com.kamelia.hedera.websocket
 
-import com.kamelia.hedera.rest.auth.SessionManager
 import com.kamelia.hedera.rest.configuration.ConfigurationEvents
 import com.kamelia.hedera.rest.configuration.GlobalConfigurationRepresentationDTO
 import com.kamelia.hedera.rest.user.UserEvents
 import com.kamelia.hedera.rest.user.UserForcefullyLoggedOutDTO
 import com.kamelia.hedera.rest.user.UserRepresentationDTO
-import com.kamelia.hedera.rest.user.UserRole
 import com.kamelia.hedera.util.defineEventListener
 import com.kamelia.hedera.util.forcefullyClose
 import com.kamelia.hedera.util.gracefullyClose
